# swagger/v1/swagger.yaml
openapi: 3.0.1
info:
  title: Hacker News
  version: 1.0.0
  description: "This is the Open API documentation for the REST API of our beloved application **Hacker News** deployed at <https://asw-hackernews-kaai12.herokuapp.com/>. <br>All operations are executable."
servers:
- url: http://18.206.152.198:8080/api
- url: http://54.158.190.88:8080/api
- url: http://3.214.224.98:8080/api
- url: http://54.175.165.61:8080/api
paths:
...
  /replies/{id}:
    get:
      tags:
        - replies
      summary: Return a reply from an ID
      operationId: showReply
      parameters:
        - name: id
          in: path
          description: ID of the reply
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reply'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                ReplyNotFound:
                  $ref: '#/components/examples/error404ReplyNotFound'
    post:
        tags:
          - replies
        summary: Create a reply for a reply
        operationId: reply_create_recursive
        parameters:
          - name: id
            in : path
            description: ID of the reply
            required: true 
            schema:
              type: integer
              format: int64
        requestBody:
          description: Attributes needed to create a reply.
          content:
            application/json:
              schema:
                required:
                - Reply
                type: object
                properties:
                  content:
                    type: string
                    example: 'This is a content example'
          required: true
        
        responses:
          201:
            description: Reply created successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Reply'
          400:
            description: 'Error: Bad Request'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  Error blank Content:
                    $ref: '#/components/examples/error400ContentBlank'
          404:
            description: 'Error: Not found'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  ReplyNotFound:
                    $ref: '#/components/examples/error404ReplyNotFound'
          401:
            description: 'Error: Unauthorized'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  NoApiKey:
                    $ref: '#/components/examples/error401'
          403:
            description: 'Error: Forbidden'
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/ErrorResult'
                examples:
                  ApiKeyNotValid:
                    $ref: '#/components/examples/error403'
        security:
        - ApiKeyAuth: []
        
    put:
      tags:
        - replies
      summary: Update a reply from an ID
      operationId: updateReply
      requestBody:
        description: Provide the new reply
        content:
          application/json:
            schema:
              required:
              - content
              type: object
              properties:
                content:
                  type: string
                  example: 'This is a content example'
        required: true
      parameters:
        - name: id
          in: path
          description: ID of the reply
          required: true
          schema:
            type: integer
            format: int64
      responses: 
        200:
          description: Reply updated
          content:
            application/json: {}
        400:
          description: 'Error: Bad Request'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                errorBlankContent:
                  $ref: '#/components/examples/error400ContentBlank'
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                ContributionNotFound:
                  $ref: '#/components/examples/error404ReplyNotFound'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                NoApiKey:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                ApiKeyNotValid:
                  $ref: '#/components/examples/error403'
                ContributionDoesNotBelong:
                  $ref: '#/components/examples/error403ContributionDoesNotBelong'

      security:
      - ApiKeyAuth: []
    
    delete:
      tags:
        - replies
      summary: Delete a reply from an ID
      operationId: deleteReply
      parameters:
        - name: id
          in: path
          description: ID of the reply
          required: true
          schema:
            type: integer
            format: int64
      responses: 
        204:
          description: Reply deleted
          content:
            application/json: {}
        404:
          description: 'Error: Not Found'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                ReplyNotFound:
                  $ref: '#/components/examples/error404ReplyNotFound'
        401:
          description: 'Error: Unauthorized'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                NoApiKey:
                  $ref: '#/components/examples/error401'
        403:
          description: 'Error: Forbidden'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResult'
              examples:
                ApiKeyNotValid:
                  $ref: '#/components/examples/error403'
                ReplyDoesNotBelong:
                  $ref: '#/components/examples/error403ReplyDoesNotBelong'

      security:
      - ApiKeyAuth: []
...
components:
  schemas:
...
    Reply:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier representing the reply
          example: 12
        content:
          type: string
          description: Content of the reply
          example: Nice comment!
        creator:
          type: string
          description: Creator of the reply
          example: cfarre
        comment_id:
          type: integer
          format: int64
          description: Unique identifier representing the comment where the reply is being posted
          example: 2
        reply_id:
          type: integer
          format: int64
          description: Unique identifier representing the parent reply where the new reply is being posted
          example: 5
        created_at:
          type: string
          format: date-time
          description: Date and time of the reply's creation
          example: "2020-11-25 14:04:56"
        updated_at:
          type: string
          format: date-time
          description: Date and time of the most recent reply's update
          example: "2020-11-25 15:04:56"
        points:
          type: integer
          format: int64
          description: Points of the reply
          example: 120
...
    ErrorResult:
      type: object
      properties:
        status:
          type: integer
          enum: [400, 401, 403, 404, 409]
        error: 
          type: string
          enum: ["Bad Request", "Unauthorized", "Forbidden", "Not Found", "Conflict"]
        message: 
          type: string
  examples:
...
    error400ContentBlank:
      value:
        status: 400
        error: "Bad Request"
        message: "Content can't be blank"
...
    error401:
      value:
        status: 401
        error: "Unauthorized"
        message: "You provided no api key (X-API-KEY Header)"
    error403:
      value:
        status: 403
        error: "Forbidden"
        message: "Your api key (X-API-KEY Header) is not valid"
    error403ContributionDoesNotBelong:
      value:
        status: 403
        error: "Forbidden"
        message: "This contribution does not belong to you"
...
    error403ReplyDoesNotBelong:
      value:
        status: 403
        error: "Forbidden"
        message: "This reply does not belong to you"
...
    error404ReplyNotFound:
      value:
        status: 404
        error : "Not Found"
        message : "No Reply with that ID"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: X-API-KEY #Custom headers in rails apps must have the X- prefix
      in: header